MDL-89015 core: Fix non-canonical scalar type casts
PHP's scalar type casting permitted some variations of scalar types. For
example, it is possible to cast a variable to an integer using both
`(integer)` and `(int)` casts. PHP scalar types have these alternate
variations:

| Type    | Canonical type name | Alternative type name |
|---------|---------------------|-----------------------|
| Integer | (int)               | (integer)             |
| Float   | (float)             | (double)              |
| Boolean | (bool)              | (boolean)             |
| String  | (string)            | (binary)              |

The alternate type names are not supported in parameter, return, or
property types. Due to this inconsistency, PHP 8.5 deprecates the four
alternative scalar type names listed above.

In PHP 9.0 and later, PHP will no longer recognize
`(integer|boolean|double|binary)` casts, and will consider them as a
cast to a class with the same name instead.

See: https://php.watch/versions/8.5/boolean-double-integer-binary-casts-deprecated

+++ public/mod/bigbluebuttonbn/classes/local/proxy/bigbluebutton_proxy.php
@@ -165,7 +165,7 @@ class bigbluebutton_proxy extends proxy_base {
             $cache->set('serverversion', $serverversion);
         }
 
-        return (double) $serverversion;
+        return (float) $serverversion;
     }
 
     /**
